fixing data type

+++ 0x14-bit_manipulation/3-set_bit.c
@@ -10,7 +10,7 @@
 
 int set_bit(unsigned long int *n, unsigned int index)
 {
-	int ibit;
+	unsigned long int ibit;
 
 	ibit = (*n & (1 << index)) ? 1 : 0;
 	if (ibit == 0)
